pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = "Telegram ecommerce bot"
authors = ["Example <example@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^20.3"
sqlmodel = "^0.0.8"
PyNaCl = "^1.5"
pydantic = "^1.10"
pyotp = "^2.9"
aiohttp = "^3.8"
monero = "^1.1.1"
python-dotenv = "^1.2.1"
bcrypt = "^4.0"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
PyJWT = "^2.8.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4"
pytest-cov = "^4.1"
pytest-asyncio = "^0.21"
ruff = "^0.0.282"
mypy = "^1.4"
httpx = "^0.26.0"

[tool.ruff]
line-length = 88

[tool.coverage.run]
branch = true
source = ["bot"]
omit = [
    "bot/__init__.py",
    "bot/handlers/__init__.py",
    "bot/services/__init__.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod"
]

[tool.pytest.ini_options]
addopts = "--cov=bot --cov-branch --cov-report=term-missing --cov-fail-under=100"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
